Revert "Flutter Driver - Create widget finders from serialized finders extensions (#67456)"
This reverts commit 74f6fa4564784090874a5023eabce4a1c8e13da0.
