+++ packages/flutter_driver/lib/src/extension/extension.dart
@@ -4,7 +4,6 @@
 
 import 'dart:async';
 
-import 'package:flutter_driver/src/common/create_finder_factory.dart';
 import 'package:meta/meta.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/foundation.dart';
@@ -124,7 +123,7 @@ class _DriverBinding extends BindingBase with SchedulerBinding, ServicesBinding,
 ///    return Some(json['title']);
 ///  }
 ///
-///  Finder createFinder(SerializableFinder finder, CreateFinderFactory finderFactory) {
+///  Finder createFinder(SerializableFinder finder) {
 ///    Some someFinder = finder as Some;
 ///
 ///    return find.byElementPredicate((Element element) {
@@ -157,13 +156,11 @@ abstract class FinderExtension {
   String get finderType;
 
   /// Deserializes the finder from JSON generated by [SerializableFinder.serialize].
-  /// [finderFactory] could be used to deserialize nested finders.
   SerializableFinder deserialize(Map<String, String> params, DeserializeFinderFactory finderFactory);
 
   /// Signature for functions that run the given finder and return the [Element]
   /// found, if any, or null otherwise.
-  /// [finderFactory] could be used to create nested finders.
-  Finder createFinder(SerializableFinder finder, CreateFinderFactory finderFactory);
+  Finder createFinder(SerializableFinder finder);
 }
 
 /// The class that manages communication between a Flutter Driver test and the
@@ -172,7 +169,7 @@ abstract class FinderExtension {
 /// This is not normally used directly. It is instantiated automatically when
 /// calling [enableFlutterDriverExtension].
 @visibleForTesting
-class FlutterDriverExtension with DeserializeFinderFactory, CreateFinderFactory {
+class FlutterDriverExtension with DeserializeFinderFactory {
   /// Creates an object to manage a Flutter Driver connection.
   FlutterDriverExtension(
     this._requestDataHandler,
@@ -357,39 +354,112 @@ class FlutterDriverExtension with DeserializeFinderFactory, CreateFinderFactory 
     return finder;
   }
 
-  @override
-  SerializableFinder deserializeFinder(Map<String, String> json) {
-    final SerializableFinder standard = super.deserializeFinder(json);
-    if (standard != null) {
-      return standard;
-    }
-
-    final String finderType = json['finderType'];
-    if (_finderExtensions.containsKey(finderType)) {
-      return _finderExtensions[finderType].deserialize(json, this);
-    }
-
-    throw DriverError('Unsupported search specification type $finderType');
+  Finder _createByTextFinder(ByText arguments) {
+    return find.text(arguments.text);
   }
 
-  @override
-  Finder createFinder(SerializableFinder finder) {
-    final Finder standard = super.createFinder(finder);
-    if(standard != null) {
-      return standard;
-    }
+  Finder _createByTooltipMessageFinder(ByTooltipMessage arguments) {
+    return find.byElementPredicate((Element element) {
+      final Widget widget = element.widget;
+      if (widget is Tooltip)
+        return widget.message == arguments.text;
+      return false;
+    }, description: 'widget with text tooltip "${arguments.text}"');
+  }
 
-    if (_finderExtensions.containsKey(finder.finderType)) {
-      return _finderExtensions[finder.finderType].createFinder(finder, this);
-    }
+  Finder _createBySemanticsLabelFinder(BySemanticsLabel arguments) {
+    return find.byElementPredicate((Element element) {
+      if (element is! RenderObjectElement) {
+        return false;
+      }
+      final String semanticsLabel = element.renderObject?.debugSemantics?.label;
+      if (semanticsLabel == null) {
+        return false;
+      }
+      final Pattern label = arguments.label;
+      return label is RegExp
+          ? label.hasMatch(semanticsLabel)
+          : label == semanticsLabel;
+    }, description: 'widget with semantic label "${arguments.label}"');
+  }
 
-    throw DriverError('Unsupported search specification type ${finder.finderType}');
+  Finder _createByValueKeyFinder(ByValueKey arguments) {
+    switch (arguments.keyValueType) {
+      case 'int':
+        return find.byKey(ValueKey<int>(arguments.keyValue as int));
+      case 'String':
+        return find.byKey(ValueKey<String>(arguments.keyValue as String));
+      default:
+        throw 'Unsupported ByValueKey type: ${arguments.keyValueType}';
+    }
+  }
+
+  Finder _createByTypeFinder(ByType arguments) {
+    return find.byElementPredicate((Element element) {
+      return element.widget.runtimeType.toString() == arguments.type;
+    }, description: 'widget with runtimeType "${arguments.type}"');
+  }
+
+  Finder _createPageBackFinder() {
+    return find.byElementPredicate((Element element) {
+      final Widget widget = element.widget;
+      if (widget is Tooltip)
+        return widget.message == 'Back';
+      if (widget is CupertinoNavigationBarBackButton)
+        return true;
+      return false;
+    }, description: 'Material or Cupertino back button');
+  }
+
+  Finder _createAncestorFinder(Ancestor arguments) {
+    final Finder finder = find.ancestor(
+      of: _createFinder(arguments.of),
+      matching: _createFinder(arguments.matching),
+      matchRoot: arguments.matchRoot,
+    );
+    return arguments.firstMatchOnly ? finder.first : finder;
+  }
+
+  Finder _createDescendantFinder(Descendant arguments) {
+    final Finder finder = find.descendant(
+      of: _createFinder(arguments.of),
+      matching: _createFinder(arguments.matching),
+      matchRoot: arguments.matchRoot,
+    );
+    return arguments.firstMatchOnly ? finder.first : finder;
+  }
+
+  Finder _createFinder(SerializableFinder finder) {
+    switch (finder.finderType) {
+      case 'ByText':
+        return _createByTextFinder(finder as ByText);
+      case 'ByTooltipMessage':
+        return _createByTooltipMessageFinder(finder as ByTooltipMessage);
+      case 'BySemanticsLabel':
+        return _createBySemanticsLabelFinder(finder as BySemanticsLabel);
+      case 'ByValueKey':
+        return _createByValueKeyFinder(finder as ByValueKey);
+      case 'ByType':
+        return _createByTypeFinder(finder as ByType);
+      case 'PageBack':
+        return _createPageBackFinder();
+      case 'Ancestor':
+        return _createAncestorFinder(finder as Ancestor);
+      case 'Descendant':
+        return _createDescendantFinder(finder as Descendant);
+      default:
+        if (_finderExtensions.containsKey(finder.finderType)) {
+          return _finderExtensions[finder.finderType].createFinder(finder);
+        } else {
+          throw 'Unsupported finder type: ${finder.finderType}';
+        }
+    }
   }
 
   Future<TapResult> _tap(Command command) async {
     final Tap tapCommand = command as Tap;
     final Finder computedFinder = await _waitForElement(
-      createFinder(tapCommand.finder).hitTestable()
+      _createFinder(tapCommand.finder).hitTestable()
     );
     await _prober.tap(computedFinder);
     return const TapResult();
@@ -397,13 +467,13 @@ class FlutterDriverExtension with DeserializeFinderFactory, CreateFinderFactory 
 
   Future<WaitForResult> _waitFor(Command command) async {
     final WaitFor waitForCommand = command as WaitFor;
-    await _waitForElement(createFinder(waitForCommand.finder));
+    await _waitForElement(_createFinder(waitForCommand.finder));
     return const WaitForResult();
   }
 
   Future<WaitForAbsentResult> _waitForAbsent(Command command) async {
     final WaitForAbsent waitForAbsentCommand = command as WaitForAbsent;
-    await _waitForAbsentElement(createFinder(waitForAbsentCommand.finder));
+    await _waitForAbsentElement(_createFinder(waitForAbsentCommand.finder));
     return const WaitForAbsentResult();
   }
 
@@ -458,7 +528,7 @@ class FlutterDriverExtension with DeserializeFinderFactory, CreateFinderFactory 
 
   Future<GetSemanticsIdResult> _getSemanticsId(Command command) async {
     final GetSemanticsId semanticsCommand = command as GetSemanticsId;
-    final Finder target = await _waitForElement(createFinder(semanticsCommand.finder));
+    final Finder target = await _waitForElement(_createFinder(semanticsCommand.finder));
     final Iterable<Element> elements = target.evaluate();
     if (elements.length > 1) {
       throw StateError('Found more than one element with the same ID: $elements');
@@ -477,7 +547,7 @@ class FlutterDriverExtension with DeserializeFinderFactory, CreateFinderFactory 
 
   Future<GetOffsetResult> _getOffset(Command command) async {
     final GetOffset getOffsetCommand = command as GetOffset;
-    final Finder finder = await _waitForElement(createFinder(getOffsetCommand.finder));
+    final Finder finder = await _waitForElement(_createFinder(getOffsetCommand.finder));
     final Element element = finder.evaluate().single;
     final RenderBox box = element.renderObject as RenderBox;
     Offset localPoint;
@@ -504,7 +574,7 @@ class FlutterDriverExtension with DeserializeFinderFactory, CreateFinderFactory 
 
   Future<DiagnosticsTreeResult> _getDiagnosticsTree(Command command) async {
     final GetDiagnosticsTree diagnosticsCommand = command as GetDiagnosticsTree;
-    final Finder finder = await _waitForElement(createFinder(diagnosticsCommand.finder));
+    final Finder finder = await _waitForElement(_createFinder(diagnosticsCommand.finder));
     final Element element = finder.evaluate().single;
     DiagnosticsNode diagnosticsNode;
     switch (diagnosticsCommand.diagnosticsType) {
@@ -523,7 +593,7 @@ class FlutterDriverExtension with DeserializeFinderFactory, CreateFinderFactory 
 
   Future<ScrollResult> _scroll(Command command) async {
     final Scroll scrollCommand = command as Scroll;
-    final Finder target = await _waitForElement(createFinder(scrollCommand.finder));
+    final Finder target = await _waitForElement(_createFinder(scrollCommand.finder));
     final int totalMoves = scrollCommand.duration.inMicroseconds * scrollCommand.frequency ~/ Duration.microsecondsPerSecond;
     final Offset delta = Offset(scrollCommand.dx, scrollCommand.dy) / totalMoves.toDouble();
     final Duration pause = scrollCommand.duration ~/ totalMoves;
@@ -544,14 +614,14 @@ class FlutterDriverExtension with DeserializeFinderFactory, CreateFinderFactory 
 
   Future<ScrollResult> _scrollIntoView(Command command) async {
     final ScrollIntoView scrollIntoViewCommand = command as ScrollIntoView;
-    final Finder target = await _waitForElement(createFinder(scrollIntoViewCommand.finder));
+    final Finder target = await _waitForElement(_createFinder(scrollIntoViewCommand.finder));
     await Scrollable.ensureVisible(target.evaluate().single, duration: const Duration(milliseconds: 100), alignment: scrollIntoViewCommand.alignment ?? 0.0);
     return const ScrollResult();
   }
 
   Future<GetTextResult> _getText(Command command) async {
     final GetText getTextCommand = command as GetText;
-    final Finder target = await _waitForElement(createFinder(getTextCommand.finder));
+    final Finder target = await _waitForElement(_createFinder(getTextCommand.finder));
 
     final Widget widget = target.evaluate().single.widget;
     String text;
